+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 * Not yet released.
 * Added detection for monolingual Gettext, XLIFF and web extension. 
 * Detect new base for Gettext and Qt TS.
+* Detect encoding of properties files.
 
 0.1
 ---